// ui/valset-standardizer-frontend/packages/src/api/generated/valset/schemas/pageResultTransferTagViewDTO.ts
/**
 * Generated by orval v6.31.0 🍺
 * Do not edit manually.
 * Java 外部估值表标准化工程，基于 Spring Boot 和 Quartz
 * OpenAPI spec version: v1.0.0
 */
import type { PageResultTransferTagViewDTOCode } from "./pageResultTransferTagViewDTOCode";
import type { TransferTagViewDTO } from "./transferTagViewDTO";

export interface PageResultTransferTagViewDTO {
  /** 返回状态码(object) */
  code?: PageResultTransferTagViewDTOCode;
  /** 返回数据 */
  data?: TransferTagViewDTO[];
  /** 返回数据格式 */
  dataType?: string;
  /** 返回消息 */
  message?: string;
  /** 当前页数 */
  pageIndex?: number;
  /** 页数 */
  pageSize?: number;
  /** 状态 */
  success?: boolean;
  /** 返回提示消息 */
  tips?: string;
  /** 总数 */
  totalCount?: number;
}